refactor ntp
Add ntp-config package to config the customized ntp.conf and ntpd.
With this change, ntp srpm is dropped to replaced by rpm.
ntp-config is configured to depend on ntp, so ntp will be included
in the ISO automatically.
ntp package will be installed in all type node, so no change to filter.

Test has been done:
build and deploy with multinode. confirm ntp.conf and ntpd is kept
the same as before.

Story: 2003768
Task: 27587

Change-Id: I795f0fd2b53c46c7302104a07c5d4cfe869d3c7b
Signed-off-by: slin14 <shuicheng.lin@intel.com>