preserve file signature when copying lighttpd config / init files
The lighttpd-config rpm was created to overwrite some configration
files from the stock lighttpd rpm.
However, the default behavior of the cp command does not
preserve extended attibutes. The result is that the IMA file
signatures were lost and the system generated IMA appraisal
failures when configured using the extended security profile.

This modification simply adds the required option to the
cp command to copy the extended attributes with the source files.

Closes-Bug: #1806433

Change-Id: I6811000b9bbd87004b371e8475359601f332a61a
Signed-off-by: Paul-Emile Element <Paul-Emile.Element@windriver.com>