Fix the lighttpd process recovery failure
When the lighttpd starts without creating a pid file, SM will restart
the process after the audit detects the failure. The init script
then kills the current running process and starts a new one.

The kill method was using the default port 80, as a result the
process that was running without a pid file was not killed. Therefore,
‘Service group web-services degraded’ alarm was never cleared due
to recovery failure.

This update changes the kill method to kill the process that running
on the configured lighttpd port.

Change-Id: Id6cae43e315d0def771b9f1422fc7753d57e6710
Closes-Bug: 1844456
Signed-off-by: Tao Liu <tao.liu@windriver.com>