Decrease tcp retransmission timeout to ~100 secs
This commit decreases the net.ipv4.tcp_retries2 to 8, which results in
a timeout of ~100 seconds down from the default of 15, with results in
a timeout of ~15 minutes.
This should result in dead connections being detected faster,
in ~100 seconds, rather than ~15 minutes, which is important when the
floating controller ip moves during a swact.

Closes-bug: 1847607
Change-Id: I48c3d396bdcd81ab11d307fd62de36deffc99f12
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>