Fix log files permissions in syslog-ng.conf
Updated /var/log files permissions to 640 from 644 to disallow
world read in order to restrict permissions of syslog-ng managed
log files but at the same time to allow group read access.
The changes are made to comply as much as possible with
openscap rules security requirements.
Verified that installation is successful for AIO-SX
and Standard 2+2 system configurations.

Story: 2008037
Task: 40694

Change-Id: I2bb2cd898bc084635e2dd158bc83cd7f2093679d
Signed-off-by: Carmen Rata <carmen.rata@windriver.com>