Controller lock failure in RR patch removal
During update strategy in a AIO-DX the swact might complete and a few
moments later a swact back be triggered by SM due to rabbitmq being able
to reach enabled state.

This happened because port 5672 was being used. No other service
should use it, suggesting that a handler was not properly disposed of in
the previous execution of rabbit. This adds a PID file parameter to
rabbit stop command of the rabbit instance, making sure that the
command will finish when the related process is terminated. Also, was
added a port check, logging when it was in use after the stop command
has concluded.

Test plan:

PASS: SX-AIO Bootstrap & deployment
PASS: DX-AIO Bootstrap & deployment

Failure path:

PASS: Launched an unexpected process to occupy port 5672 and it is
logged successfully

Closes-bug: #1942464
Signed-off-by: Iago Regiani <Iago.RodriguezRegiani@windriver.com>
Change-Id: Id6c0770ba8e1a5460e4af07e0e6d0d5447581771