tuned: request performance cpu governor
Problem:
StarlingX uses the balanced tuned profile. This is not exactly
what we want as the balanced profile attempts to set the cpu
governor to conservative. With the 3.10 kernel and drivers in
place, we were getting away with this as the request to set the
governor to conservative fails because the scaling driver (intel_pstate)
doesn't support it. This failure leaves the performance governor
in place, which is the cpu governor that we actually want.

With the 5.10 kernel and drivers, the scaling driver (intel_cpufreq)
does support conservative, so we end up with that governor which
trades off performance for power savings, instead of the desired
performance governor.

Solution:
Create a new starlingx tuned profile from the balanced profile,
but change its governor setting to performance instead of
conservative.  This leaves the tuned behavior exactly as it was
before.  This new starlingx tuned profile gives us greater flexibility
as we can adjust it in the future to suit our specific needs.

We also introduce a new recommend.conf to select the new starlingx
profile.  We give this new file rpm type noreplace.  This
allows the end user to change it to select a different profile
without us overwriting their change on a package upgrade.  The
starlingx profile config however we will potentially overwrite on
an upgrade as an end user shouldn't be modifying it.  They
should create their own profile if they feel such a need.

Change-Id: Ib249aa63326a4018e1accfdef8b622dcad9ab03d
Closes-Bug: 1944392
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>