Modified io-scheduler udev rules with 5.10 kernel
5.10 kernel introduced four io-schedulers 'mq-deadline' 'kyber' 'bfq'
and 'none'.

The 'none' option is roughly comparable to 'noop', 'mq-deadline' is
comparable to 'deadline', and 'bfq' is comparable to 'cfq'. So we
change 'cfq' to 'bfq', 'noop' to 'none', 'deadline' to 'mq-deadline'.
and the settings aligned as much as possible with what we were using
for the 3.10 kernel.

Verification:
Successfully installed lowlat (rt) on AIO-SX and std on a 2 controller/2
compute system.  Nothing was seen to indicate that i/o was too slow or
causing timeouts.

Closes-Bug: #1947551

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: Ie833ff3972f7f0c67027015ad5107d1c2e66c166