Link nfsX to nfs-X service for nfscommon and nfsserver
This work is part of Debian integration effort.
This work only affects Debian.

The services used on CentOS are named nfscommon and nfsserver.
To keep compatibility with existing service dependencies and
ocf scripts, link nfscommon to nfs-common and nfsserver to nfs-server,
so they are used interchangeable.
We don't have to worry about mask/umask operations as these are not
used.

Maybe in a future work drop the nfscommon and nfsserver naming when
CentOS support is dropped, we would have to update services and ocf
scripts.

Tests on AIO-SX:
PASS: build-image for standard kernel
PASS: unlocked enabled available
PASS: nfscommon points to nfs-common,
      nfsserver points to nfs-server

Story: 2009101
Task: 44778
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I7359d440e1585f793ba4e78670ef5cf7a7a02960