Killing "sshd" does not trigger an alarm event
It was identified that the sshd service was being restarted
automatically by systemd because the ssh.service file had the
"Restart=on-failure" set, which prevented pmon to proper monitor it
resulting in the alarm not being set.
Comparing with CentOS, we confirmed that this line in the service
file was commented out.
However, checking the openssh-config for Debian, the line was also
commented out, but the file, sshd.service is not the file used by
openssh-server package in Debian, instead, it is the ssh.service.
This commit replaces sshd.service file by ssh.service (based on the
file from the Debian package openssh-server) and add the modification
to comment out the "Restart=on-failure".

Test Plan
PASS: Build and install
PASS: Unlock AIO-SX
PASS: Verified that killing the sshd process 10 times would
trigger the alarm

Closes-bug: 1991400

Signed-off-by: Manasses Julio <manasses.dasilvajunior@windriver.com>
Change-Id: I477d5261ccf44ccb29c7b3ab0dbc9bd816a753cc