After upgrade subcloud linux prompt is different
The regex in prompt.sh is incorrectly adding [] to around
user@hostname. This commit removes those brackets in order
to have the expected format.

Test Plan:

PASS: Enter "sudo su -" and verify if the prompt shows
            "root@controller-0:~$".
PASS: Generate a new .iso with the changes and verify
      if when entering sudo with "sudo su -" it is showing
      "root@controller-0:~$".
PASS: Install system and verify if prompt show
      "sysadmin@controller-0:~$ "

Closes-Bug: 1997972
Related-Bug: 1995988
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: Idfd3b03f2ce1eb9679e227c9406aaf200514977a