Debian: Save the log with a local timestamp
A lot of applications get the timezone information from
the cache, so they won't be aware of the timezone's
changing, and the log's timestamp does not change with
the timezone changing.
Use the local time as a timestamp for the log instead
of the timestamp of the sender's messages can fix it.

Test Plan:
PASS: Change the timezone and reload the syslog-ng
      service ,the timestamp of the log is equal to
      local time.

Closes-bug: 1997930
Signed-off-by: Wentao Zhang<Wentao.Zhang@windriver.com>
Change-Id: I13ec19d4f78ea3d5f6a7c20873c082add5fe412e