Delete rabbit_stop from rabbit_start
On DC with central cloud as standard, sometimes occurs
that RabbitMQ doesn't start properly and keeps being restarted
by sm, either on system restart or restarting RabbitMQ service.
It was found that part of the problem was the presence of
the rabbit_stop function inside the rabbit_start,
which was added in the past because in a previous version
of RabbitMQ the service didn't completely stop before the
call of the start function. Since this problem doesn't occur
on Debian, the rabbit_stop call was removed, and along side
[1], fixes this problem.

Test Plan:
PASS - Run sm-restart service rabbit successfully check rabbit was
  running as expected. This test was used to recreate the bug.
PASS - Reboot the host successfully and check rabbit was running as
  expected.
PASS - Lock/unlock and check if rabbit was running as expected.

Regression test:
PASS - Install and bootstrap DC
PASS - Install and bootstrap DX

[1]
Depends-on: https://review.opendev.org/c/starlingx/ha/+/866208

Closes-bug: 1997966

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: Ibd49a83f58a7f7efcf7b5c026aaf3e86d985e31d