use new location for k8s-container-cleanup.sh
The debian packaging guidelines do not want packages to install
binaries under /usr/local so we're moving the k8s-container-cleanup.sh
executable to /usr/sbin.

This should be merged with the corresponding change in the "integ"
repo, but it's not the end of the world if it isn't done at exactly
the same time.

Test Plan:
PASS Install AIO-SX with K8s 1.25 and upgrade to K8s 1.26.
PASS Install with K8s 1.26 on AIO-SX and ensure the system comes up
     normally.

Story: 2010368
Task: 48012

Change-Id: I4a536d27364df2c57a8c375684d14929b45f13fb
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>