Remove unused dhcp client options
With the IPSec feature DHCP clients in starlingx no longer need to
request the options: domain-name-servers, and
dhcp6.name-servers.

The reason is that the management network is been migrated to a static
address configuration, with gateway and DNS servers provided to
the config files (ifcfg-* and resolv.conf) first by kickstart and
after by puppet manifest configuration.

The pxeboot network will use DHCP for initial bringup but it does not
need those options.

Test Plan:
[PASS] Install IPv4 Standard with storage
[PASS] Install IPv6 Standard with storage

Story: 2010940
Task: 49385
Depends-On: https://review.opendev.org/c/starlingx/config/+/904000

Change-Id: I2dc8f44aaf9d22881ae24679795f14b4e5b648d1
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>