Improve DNSMasq init script
During the first unlock after a fresh install, the DNSMasq is started
but we noticed that other services that depends on it are not
receiving the DNS query answers in time, as a result the services are
restarted by SM and everything works fine.
This issue doesn't happen for the next lock/unlocks.
To mitigate this issue, the DNSMasq start process will have 20
seconds to resolve the "controller.internal", so the next services
handled by SM will start just when DNSMasq is fully operational.

Test done:
AIO-SX fresh install
AIO-DX fresh install
AIO-DX host-swact

Story: 2010722
Task: 50243

Depends-On: https://review.opendev.org/c/starlingx/config/+/920694


Change-Id: I405de2248a8b834cbc297c3e550566052264f343
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>