Fix enable rsync for /etc/update.d
When it is needed to do a /etc/update.d rsync between controllers,
rsync does not find this folder.

This commit adds /etc/update.d directory to rsync scope for Debian.

This is an addendum to the following commit:
Relates-to: I2a68091704518a4cae8507351c23e4a84d1ea027

Test-plan:
PASS: rsync a /etc/update.d content between controllers

Story: 2010676
Task: 51299

Change-Id: I0a33c8a97cb27a3bb9bee6acfbc6f72d71292580
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>