Modifying syslog-ng.conf to account for sm.conf
Introduced a new file sm.conf as part of Story: 2010533
Hence moving SM related syslog configuration to the new file.

Depends-On:https://review.opendev.org/c/starlingx/ha/+/940153

Test Plan:

AIO-Duplex : Verified that the changes are available during the initial host installation with the ISO.
AIO-Duplex:  Verified that the changes are seen during an upgrade.
AIO-Duplex: Verified the changes are available during a swact and subsequent host lock and unlock

Closes-Bug: 2083632
Change-Id: I3dda49d041514b86b3841feb8ffc2089d96100a2
Signed-off-by: Sandhya Kalisetty sandhya.kalisetty@windriver.com